backend/requirements.txt
# Django Framework
django>=5.0,<5.1
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<4.4

# WebSocket support (Channels for Django)
channels>=4.0,<4.1
channels-redis>=4.1,<4.2

# Database drivers
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<2.2

# Environment and configuration
python-dotenv>=1.0,<2.0

# Server (for production with ASGI)
gunicorn>=21.2,<22.0
uvicorn>=0.25,<0.26
daphne>=4.0,<5.0

# Security and Auth
pyjwt>=2.8,<3.0
cryptography>=41.0,<42.0
requests>=2.31,<3.0


# AWS S3
boto3==1.34.0
django-storages==1.14.0

dj-rest-auth==5.0.1
django-allauth==0.58.0
djangorestframework-simplejwt==5.3.0